requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
cors==1.0.1
dataclasses-json==0.6.3
dataclasses-json-speakeasy==0.5.11
Deprecated==1.2.14
distlib==0.3.8
distro==1.8.0
dnspython==2.4.2
emoji==2.9.0
exceptiongroup==1.2.0
fastapi==0.105.0
filelock==3.13.1
filetype==1.2.0
Flask==3.0.0
Flask-Cors==4.0.0
flatbuffers==23.5.26
frozenlist==1.4.1
fsspec==2023.12.2
future==0.18.3
gevent==23.9.1
gitdb==4.0.11
GitPython==3.1.40
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-generativeai==0.3.2
google_search_results==2.4.2
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio
grpcio-status
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.20.1
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kubernetes==28.1.0
langchain==0.0.352
langchain-community==0.0.6
langchain-core==0.1.3
langchain-google-genai==0.0.6
langdetect==1.0.9
langsmith==0.0.75
loguru==0.7.2
lxml==4.9.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
mmh3==4.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numexpr==2.8.8
numpy==1.26.2
oauthlib==3.2.2
onnxruntime==1.16.3
openai==1.6.1
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
overrides==7.4.0
packaging==23.2
pandas==2.1.4
pdf2image==1.16.3
Pillow==10.1.0
pinecone-client==2.2.4
platformdirs==4.2.0
posthog==3.1.0
proto-plus==1.23.0
protobuf==4.25.1
pulsar-client==3.3.0
pyarrow==14.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.1
PyPDF2==3.0.1
PyPika==0.48.9
pyreadline3==3.4.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-iso639==2023.12.11
python-magic==0.4.27
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==3.5.2
referencing==0.32.0
regex==2023.10.3
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.15.2
rsa==4.9
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
starlette==0.27.0
streamlit==1.29.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.5.2
tldextract==5.1.1
tokenizers==0.15.0
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
unstructured==0.11.6
unstructured-client==0.15.1
uritemplate==4.1.1
urllib3==1.26.18
uvicorn==0.25.0
validators==0.22.0
virtualenv==20.25.0
watchdog==3.0.0
watchfiles==0.21.0
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.1
wikipedia==1.4.0
win32-setctime==1.1.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
pip~=24.0
argparse~=1.4.0
wheel~=0.41.2
setuptools~=65.5.1
psutil~=5.9.8
torch~=2.2.0
uvloop~=0.19.0
transformers~=4.37.2
